refactor(app): extract useFileSession hook
